% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_ass_content}
\alias{generate_ass_content}
\title{Generate ASS subtitle file content for fish tracking}
\usage{
generate_ass_content(
  df,
  width,
  height,
  frame_rate,
  trail_length = 10,
  start_size = 30,
  size_step = 3,
  fish_colors = NULL
)
}
\arguments{
\item{df}{Data frame containing fish tracking data}

\item{width}{Video width in pixels}

\item{height}{Video height in pixels}

\item{frame_rate}{Video frame rate}

\item{trail_length}{Number of previous positions to show}

\item{start_size}{Starting size for most recent position}

\item{size_step}{Size reduction per previous position}

\item{fish_colors}{Vector of color codes in RRGGBB format}
}
\value{
Character vector containing ASS subtitle content
}
\description{
Generate ASS subtitle file content for fish tracking
}
\keyword{internal}
