% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_video_resolution}
\alias{get_video_resolution}
\title{Get video resolution using ffprobe}
\usage{
get_video_resolution(video_path)
}
\arguments{
\item{video_path}{Character string of the video file path}
}
\value{
List containing width and height of the video
}
\description{
Get video resolution using ffprobe
}
