% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{seconds_to_ass_time}
\alias{seconds_to_ass_time}
\title{Convert Seconds to ASS Time Format}
\usage{
seconds_to_ass_time(t)
}
\arguments{
\item{t}{Time in seconds}
}
\value{
Character string in ASS format (H:MM:SS.CS)
}
\description{
Convert Seconds to ASS Time Format
}
\keyword{internal}
