% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{update_trail}
\alias{update_trail}
\title{Update a position trail with a new position}
\usage{
update_trail(trail, new_position, max_length)
}
\arguments{
\item{trail}{List of position coordinates}

\item{new_position}{List containing x and y coordinates}

\item{max_length}{Integer specifying maximum length of the trail}
}
\value{
Updated trail list
}
\description{
Update a position trail with a new position
}
\keyword{internal}
