% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_video_resolution}
\alias{get_video_resolution}
\title{Get Video Resolution Using ffprobe}
\usage{
get_video_resolution(video_path)
}
\arguments{
\item{video_path}{Path to video file}
}
\value{
List with width and height
}
\description{
Get Video Resolution Using ffprobe
}
