% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{seconds_to_ass_time}
\alias{seconds_to_ass_time}
\title{Convert seconds to ASS time format}
\usage{
seconds_to_ass_time(t)
}
\arguments{
\item{t}{Numeric value representing time in seconds}
}
\value{
Character string in ASS time format (H:MM:SS.FF)
}
\description{
Convert seconds to ASS time format
}
